Improper use of Intersection Type
See https://www.typescriptlang.org/docs/handbook/unions-and-intersections.html#intersection-types.

Using `&` might be convenient while defining types, but leads to ugly situations when using theses types. With Union Types(`|`) TypeScript would correctly infer the actual type by using a simple if or and type guard. But like in the given example everything is merged together and there is no way to differentiate between a entity or context etc. Which is forcing the consumer to do a type cast.